# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Sidekeek API Service",
    "description": "Group of services provided to maintain sidekeek-client needs",
    "contact": "yourname@something.com",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/sideAPIkeek",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/editAccount": {
      "post": {
        "tags": ["ProfileController"],
        "summary": "edit account",
        "description": "edit user account",
        "operationId": "editAccount",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "sessionCode",
            "in" : "formData",
            "description": "sessionCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "companyName",
            "in" : "formData",
            "description": "companyName",
            "required": true,
            "type": "string"
          },
          {
            "name": "about",
            "in" : "formData",
            "description": "about",
            "required": true,
            "type": "string"
          },
          {
            "name": "handphone",
            "in" : "formData",
            "description": "handphone",
            "required": true,
            "type": "string"
          },
          {
            "name": "city",
            "in" : "formData",
            "description": "city",
            "required": true,
            "type": "string"
          },
          {
            "name": "address",
            "in" : "formData",
            "description": "address",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/editAccount"
            }
          }
        }
      }
    },
...
    "/confirmation": {
      "post": {
        "tags": ["RegisterController"],
        "summary": "confirm code",
        "description": "confirm code to validate email",
        "operationId": "confirmation",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "uniqueCode",
            "in" : "formData",
            "description": "unique_code",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/confirmation"
            }
          }
        }
      }
    },
...
    "/hostSignUp": {
      "post": {
        "tags": ["RegisterController"],
        "summary": "sign up new host",
        "description": "sign up new user for host",
        "operationId": "hostSignUp",
        "parameters": [
          {
            "name": "token",
            "in" : "formData",
            "description": "token",
            "required": true,
            "type": "string"
          },
          {
            "name": "idTipe",
            "in" : "formData",
            "description": "id tipe",
            "required": true,
            "type": "string"
          },
          {
            "name": "idCat",
            "in" : "formData",
            "description": "id category",
            "required": true,
            "type": "string"
          },
          {
            "name": "compName",
            "in" : "formData",
            "description": "company name",
            "required": true,
            "type": "string"
          },
          {
            "name": "threadTitle",
            "in" : "formData",
            "description": "thread title",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in" : "formData",
            "description": "email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in" : "formData",
            "description": "password",
            "required": true,
            "type": "string"
          },
          {
            "name": "confirmation",
            "in" : "formData",
            "description": "confirmation",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref" : "#/definitions/hostSignUp"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "hostSignUp": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        },
        "status": {
          "example": ""
        },
        "unique_code": {
          "example": ""
        },
        "email": {
          "example": ""
        }
      }
    },
...
    "confirmation": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        },
        "session": {
          "example": ""
        },
        "email": {
          "example": ""
        }
      }
    },
...
    "editAccount": {
      "properties": {
        "message": {
          "example": ""
        },
        "error": {
          "example": ""
        }
      }
    },
...
  }
}